- Organized classes into packages - Added an empty Weapon class so the weapon package would push
