+++ ZombieWar/src/main/java/com/csc422/zombiewar/characters/CommonInfect.java
@@ -1,4 +1,4 @@
-package com.csc422.zombiewar;
+package com.csc422.zombiewar.characters;
 
 public class CommonInfect extends Zombie
 {
