+++ ZombieWar/src/main/java/com/csc422/zombiewar/weapons/Weapon.java
@@ -1,0 +1,14 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.csc422.zombiewar.weapons;
+
+/**
+ *
+ * @author caitl
+ */
+public class Weapon {
+    
+}
